KNL-205 Support for files > 2G in Content Hosting
git-svn-id: https://source.sakaiproject.org/svn/kernel/trunk@65808 66ffb92e-73f9-0310-93c1-f5514f145a0a

+++ kernel/api/src/main/java/org/sakaiproject/content/api/ContentResource.java
@@ -38,7 +38,7 @@ public interface ContentResource
 	* Access the content byte length.
 	* @return The content byte length.
 	*/
-	public int getContentLength();
+	public long getContentLength();
 
 	/**
 	* Access the resource MIME type.
